app: refactor folders name
