--- src/main/kotlin/com/example/shorturl/ShortUrlApplication.kt
@@ -1,11 +1,0 @@
-package com.example.shorturl
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class ShortUrlApplication
-
-fun main(args: Array<String>) {
-    runApplication<ShortUrlApplication>(*args)
-}
