--- src/main/kotlin/com/example/shorturl/datasource/DataSourceTestConfig.kt
@@ -1,20 +1,0 @@
-package com.example.shorturl.datasource
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Profile
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType
-import javax.sql.DataSource
-
-@Configuration
-@Profile("test")
-class DataSourceTestConfig {
-    @Bean
-    fun dataSource(): DataSource {
-        return EmbeddedDatabaseBuilder()
-            .setType(EmbeddedDatabaseType.H2)
-            .addScript("schema.sql")
-            .build()
-    }
-}
