--- src/main/kotlin/com/example/shorturl/datasource/service/UrlService.kt
@@ -1,13 +1,0 @@
-package com.example.shorturl.datasource.service
-
-import com.example.shorturl.datasource.Url
-import com.example.shorturl.datasource.repository.UrlRepository
-import org.springframework.stereotype.Service
-
-@Service
-class UrlService(private val db: UrlRepository) {
-    fun save(url: Url): Url = db.save(url)
-    fun findByUrl(urlPath: String): Url? = db.findByGeneratedUrl(urlPath)
-    fun findExpiredUrls(): List<Url> = db.findExpiredUrls()
-    fun delete(url: Url) = db.delete(url)
-}
