+++ src/main/kotlin/me/michelao/shorturl/ShortUrlApplication.kt
@@ -1,0 +1,11 @@
+package me.michelao.shorturl
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class ShortUrlApplication
+
+fun main(args: Array<String>) {
+    runApplication<ShortUrlApplication>(*args)
+}
