+++ src/main/kotlin/me/michelao/shorturl/configuration/properties/MyAppProperties.kt
@@ -1,0 +1,16 @@
+package me.michelao.shorturl.configuration.properties
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@ConfigurationProperties("app")
+class MyAppProperties {
+    lateinit var baseUrl: String
+    lateinit var bucketName: String
+    val expiration = Expiration()
+    class Expiration {
+        var minDays: Int = 30
+        var maxDays: Int = 365
+    }
+}
