+++ src/main/kotlin/me/michelao/shorturl/controller/FileUploadController.kt
@@ -1,0 +1,93 @@
+package me.michelao.shorturl.controller
+
+import me.michelao.shorturl.UrlHandler
+import me.michelao.shorturl.configuration.properties.MyAppProperties
+import me.michelao.shorturl.datasource.MimeType
+import me.michelao.shorturl.datasource.S3ClientData
+import me.michelao.shorturl.datasource.Url
+import me.michelao.shorturl.datasource.service.UrlService
+import me.michelao.shorturl.tools.ExpirationCalculator
+import jakarta.servlet.http.HttpServletRequest
+import org.springframework.http.MediaType
+import org.springframework.http.ResponseEntity
+import org.springframework.stereotype.Controller
+import org.springframework.web.bind.annotation.CrossOrigin
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestHeader
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.multipart.MultipartFile
+import java.util.Date
+import java.util.logging.Logger
+import kotlin.math.pow
+
+@Controller
+class FileUploadController(
+    private val service: UrlService,
+    private val s3: S3ClientData,
+    private val appProperties: MyAppProperties,
+    private val expirationCalculator: ExpirationCalculator,
+) {
+    private val logger = Logger.getLogger(this.javaClass.name)
+    fun readRealIp(request: HttpServletRequest): String {
+        var ip: String? = request.getHeader("x-forwarded-for")
+        if (ip == null) {
+            ip = request.getHeader("x-real-ip")
+        }
+        if (ip == null) {
+            ip = request.remoteAddr
+        }
+        return ip
+    }
+
+    @PostMapping("/", produces = [MediaType.TEXT_PLAIN_VALUE])
+    @CrossOrigin(exposedHeaders = ["X-Token"])
+    @ResponseBody
+    fun postFile(
+        @RequestHeader("User-Agent") userAgent: String?,
+        request: HttpServletRequest,
+        @RequestParam("file") file: MultipartFile,
+        @RequestParam("expires") expires: Date? = null
+    ): Any {
+        if (!file.isEmpty) {
+            val contentType = file.contentType ?: MediaType.APPLICATION_OCTET_STREAM.toString()
+            val originalFileName = file.originalFilename
+            val sizeMebibytes = file.size / 2.0.pow(20)
+            var generatedUrl: String
+            do {
+                generatedUrl = UrlHandler.generatorURL(14)
+                val findUrl = service.findByUrl("/$generatedUrl")
+            } while (findUrl != null)
+            var expiresConverted = expirationCalculator.calculateExpiresAt(sizeMebibytes)
+            if (expires != null) {
+                if (expirationCalculator.testExpireUnderLimit(expires)) {
+                    expiresConverted = expires
+                } else {
+                    return ResponseEntity.badRequest().body("The expiry date has exceeded the maximum age")
+                }
+            }
+            val realIp = readRealIp(request)
+            val token = UrlHandler.generatorToken()
+            logger.info("IP: $realIp, Upload filename: $originalFileName")
+            val url = Url(
+                originalFileName,
+                "/$generatedUrl",
+                MimeType.fromValue(contentType),
+                sizeMebibytes.toLong(),
+                token,
+                realIp,
+                userAgent,
+                expiresConverted
+            )
+            service.save(url)
+            s3.writeData(generatedUrl, file.inputStream, contentType)
+            return ResponseEntity
+                .ok()
+                .headers {
+                    it.set("X-Token", token)
+                }
+                .body("${appProperties.baseUrl}/${generatedUrl}")
+        }
+        return ResponseEntity.badRequest().body("File to upload cannot be empty")
+    }
+}
