+++ src/main/kotlin/me/michelao/shorturl/datasource/DataSourceTestConfig.kt
@@ -1,0 +1,20 @@
+package me.michelao.shorturl.datasource
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType
+import javax.sql.DataSource
+
+@Configuration
+@Profile("test")
+class DataSourceTestConfig {
+    @Bean
+    fun dataSource(): DataSource {
+        return EmbeddedDatabaseBuilder()
+            .setType(EmbeddedDatabaseType.H2)
+            .addScript("schema.sql")
+            .build()
+    }
+}
