+++ src/test/kotlin/me/michelao/shorturl/FileUploadTests.kt
@@ -1,0 +1,57 @@
+package me.michelao.shorturl
+
+import me.michelao.shorturl.configuration.properties.MyAppProperties
+import jakarta.servlet.ServletContext
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.mock.web.MockMultipartFile
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
+import java.util.concurrent.TimeUnit
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.ANY)
+@AutoConfigureMockMvc
+@ActiveProfiles("test")
+class FileUploadTests(@Autowired private val myApp: MyAppProperties) {
+    @Autowired
+    private lateinit var servletContext: ServletContext
+    @Test
+    fun testUpload(@Autowired mvc: MockMvc) {
+        val file = MockMultipartFile(
+            "file",
+            "fileTest.txt",
+            "text/plain",
+            "Hello world".toByteArray()
+        )
+        mvc.perform {
+            multipart("/")
+                .file(file)
+                .buildRequest(servletContext)
+        }
+            .andExpect(status().isOk)
+    }
+
+    @Test
+    fun testUploadExceedMaxAge(@Autowired mockMvc: MockMvc) {
+        val file = MockMultipartFile(
+            "file",
+            "fileTest.txt",
+            "text/plain",
+            "Hello world".toByteArray()
+        )
+        val expireTime = TimeUnit.DAYS.toSeconds((myApp.expiration.maxDays + 1).toLong())
+        mockMvc.perform {
+            multipart("/")
+                .file(file)
+                .param("expires", "${expireTime}s")
+                .buildRequest(servletContext)
+        }
+            .andExpect(status().is4xxClientError)
+    }
+}
